requirements.txt
Flask==1.1.1
requests==2.22.0
pymongo==3.10.1
python-dotenv==0.13.0
